requirements.txt
# Python dependencies for kickstarter-pipeline
# Install with: pip install -r requirements.txt

# Data handling
pandas>=2.0.0
pyarrow>=13.0.0

# GCP
google-cloud-storage>=2.12.0
google-cloud-bigquery>=3.12.0
google-cloud-bigquery[pandas]>=3.12.0
google-auth>=2.23.0

# HTTP
requests>=2.31.0

# dbt
dbt-bigquery>=1.7.0
dbt-core>=1.7.0